test(e2e): PR-run the wiring-guard policy; reject a pending peer=
Finding 2: move the synthetic-item wiring-guard tests (pure _wiring_error policy,
incl. the peer branches) to tests/framework_conformance/test_agent_wiring_rules.py
so they run on every PR; the e2e-tree guards file kept only the pytester
real-collection (fixture-closure) cases, which the E2E gate is fine for. Closes the
gap where the wiring policy (and the new peer branches) had no PR coverage.

Finding 3: per_adapter(peer=...) now rejects a registered-but-pending adapter at
decoration time (pending adapters run no cells), alongside the not-registered check.
